Uitwerking van de workshop

+++ src/main/java/han/aim/se/noyoumaynot/movie/controller/MovieController.java
@@ -1,17 +1,20 @@
 package han.aim.se.noyoumaynot.movie.controller;
 
 import han.aim.se.noyoumaynot.movie.domain.Movie;
+import han.aim.se.noyoumaynot.movie.domain.Role;
+import han.aim.se.noyoumaynot.movie.domain.User;
 import han.aim.se.noyoumaynot.movie.service.AuthenticationService;
 import han.aim.se.noyoumaynot.movie.service.MovieService;
-import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
+import java.util.List;
 
 @RestController
-@RequestMapping("/movies")
+@RequestMapping("/")
 public class MovieController {
     private final MovieService movieService;
     private final AuthenticationService authenticationService;
@@ -22,37 +25,49 @@ public class MovieController {
         this.authenticationService = authenticationService;
     }
 
-    @GetMapping
-    public ArrayList<Movie> getAllMovies() {
+    @PostMapping("/login")
+    public String login(@RequestBody User user) {
+        return authenticationService.login(user.getUsername(), user.getPassword()).getToken();
+    }
+
+    @GetMapping("/movies")
+    public List<Movie> getAllMovies(@RequestHeader(value = HttpHeaders.AUTHORIZATION, required = false) String token) {
+        authenticate(token);
         return movieService.getMovieList();
     }
 
-    @GetMapping("/show")
-    public Movie getMovieById(@RequestParam("id") String id) {
-
-            Movie movie = movieService.getMovieById(id);
-            return movie;
+    @GetMapping("/movies/show")
+    public Movie getMovieById(@RequestParam("id") String id, @RequestHeader(value = HttpHeaders.AUTHORIZATION, required = false) String token) {
+        authenticate(token);
+        Movie movie = movieService.getMovieById(id);
+        return movie;
     }
 
-    @PostMapping("/add")
-    public Movie addMovie(@RequestBody Movie movie) {
+    @PostMapping("/movies/add")
+    public Movie addMovie(@RequestBody Movie movie, @RequestHeader(value = HttpHeaders.AUTHORIZATION, required = false) String token) {
+        Role role = authenticate(token);
+        if (!role.getIsAdmin()) {
+            throw new AuthorizationException("Unauthorized");
+        }
         movieService.insertMovie(movie);
         return movie;
     }
 
-    @DeleteMapping("/delete/{id}")
-    public ResponseEntity<String> deleteMovie(@PathVariable("id") String id) {
+    @DeleteMapping("/movies/delete/{id}")
+    public ResponseEntity<String> deleteMovie(@PathVariable("id") String id, @RequestHeader(value = HttpHeaders.AUTHORIZATION, required = false) String token) {
+        Role role = authenticate(token);
+        if (!role.getIsAdmin()) {
+            throw new AuthorizationException("Unauthorized");
+        }
         movieService.deleteMovie(id);
         return ResponseEntity.ok().build();
     }
 
-    private String authenticate(HttpServletRequest request) throws Exception {
-        if (authenticationService.isValidToken("")){
-            return authenticationService.getUsername("");
+    private Role authenticate(String token) throws AuthenticationException {
+        if (authenticationService.isValidToken(token)) {
+            return authenticationService.getRole(token);
         } else {
-            throw new Exception("Invalid token");
+            throw new AuthenticationException("Invalid token");
         }
     }
-
-
 }
